Split Hyprland config format helpers
