+++ internal/config/hyprland.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
 	"os"
@@ -154,254 +153,6 @@ func VerifyIncludeChain(format HyprConfigFormat, rootConfigPath string, targetPa
 		return VerifyLuaIncludeChain(rootConfigPath, targetPath)
 	}
 	return VerifySourceChain(rootConfigPath, targetPath)
-}
-
-func VerifyLuaIncludeChain(rootConfigPath string, targetPath string) error {
-	rootConfigPath, err := resolvePath(rootConfigPath, "")
-	if err != nil {
-		return err
-	}
-	targetPath, err = resolvePath(targetPath, "")
-	if err != nil {
-		return err
-	}
-
-	ok, err := isLuaPathIncluded(rootConfigPath, targetPath, map[string]bool{})
-	if err != nil {
-		return err
-	}
-	if ok {
-		return nil
-	}
-
-	return fmt.Errorf("%s is not included by %s; add `dofile(%q)` to your Hyprland Lua config or pass a different --monitors-conf target", targetPath, rootConfigPath, targetPath)
-}
-
-func isLuaPathIncluded(rootConfigPath string, targetPath string, visited map[string]bool) (bool, error) {
-	rootConfigPath = filepath.Clean(rootConfigPath)
-	targetPath = filepath.Clean(targetPath)
-	if rootConfigPath == targetPath {
-		return true, nil
-	}
-	if visited[rootConfigPath] {
-		return false, nil
-	}
-	visited[rootConfigPath] = true
-
-	content, err := os.ReadFile(rootConfigPath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return false, fmt.Errorf("Hyprland config %s does not exist", rootConfigPath)
-		}
-		return false, err
-	}
-
-	for _, includeValue := range parseLuaIncludeCalls(string(content)) {
-		includePath, err := resolvePath(includeValue, filepath.Dir(rootConfigPath))
-		if err != nil {
-			return false, err
-		}
-		if includePath == targetPath {
-			return true, nil
-		}
-		info, err := os.Stat(includePath)
-		if err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				continue
-			}
-			return false, err
-		}
-		if info.IsDir() {
-			continue
-		}
-		ok, err := isLuaPathIncluded(includePath, targetPath, visited)
-		if err != nil {
-			return false, err
-		}
-		if ok {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func parseLuaIncludeCalls(content string) []string {
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	out := make([]string, 0)
-	for scanner.Scan() {
-		line := stripLuaComments(scanner.Text())
-		out = append(out, parseLuaLiteralCall(line, "dofile")...)
-		out = append(out, parseLuaLiteralCall(line, "source")...)
-	}
-	return out
-}
-
-func stripLuaComments(line string) string {
-	if idx := strings.Index(line, "--"); idx >= 0 {
-		return line[:idx]
-	}
-	return line
-}
-
-func parseLuaLiteralCall(line string, name string) []string {
-	line = strings.TrimSpace(line)
-	out := make([]string, 0, 1)
-	for {
-		idx := strings.Index(line, name)
-		if idx < 0 {
-			return out
-		}
-		beforeOK := idx == 0 || !isLuaIdentifierByte(line[idx-1])
-		line = line[idx+len(name):]
-		if !beforeOK {
-			continue
-		}
-		line = strings.TrimLeft(line, " \t")
-		if !strings.HasPrefix(line, "(") {
-			continue
-		}
-		line = strings.TrimLeft(line[1:], " \t")
-		if line == "" || (line[0] != '\'' && line[0] != '"') {
-			continue
-		}
-		quote := line[0]
-		line = line[1:]
-		end := strings.IndexByte(line, quote)
-		if end < 0 {
-			continue
-		}
-		value := line[:end]
-		line = line[end+1:]
-		if strings.TrimSpace(value) != "" {
-			out = append(out, value)
-		}
-	}
-}
-
-func isLuaIdentifierByte(b byte) bool {
-	return b == '_' || b >= '0' && b <= '9' || b >= 'A' && b <= 'Z' || b >= 'a' && b <= 'z'
-}
-
-func VerifySourceChain(rootConfigPath string, targetPath string) error {
-	rootConfigPath, err := resolvePath(rootConfigPath, "")
-	if err != nil {
-		return err
-	}
-	targetPath, err = resolvePath(targetPath, "")
-	if err != nil {
-		return err
-	}
-
-	ok, err := isPathSourced(rootConfigPath, targetPath, map[string]bool{})
-	if err != nil {
-		return err
-	}
-	if ok {
-		return nil
-	}
-
-	return fmt.Errorf("%s is not sourced by %s; add `source = %s` to your Hyprland config or pass a different --monitors-conf target", targetPath, rootConfigPath, targetPath)
-}
-
-func isPathSourced(rootConfigPath string, targetPath string, visited map[string]bool) (bool, error) {
-	rootConfigPath = filepath.Clean(rootConfigPath)
-	targetPath = filepath.Clean(targetPath)
-	if rootConfigPath == targetPath {
-		return true, nil
-	}
-	if visited[rootConfigPath] {
-		return false, nil
-	}
-	visited[rootConfigPath] = true
-
-	content, err := os.ReadFile(rootConfigPath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return false, fmt.Errorf("Hyprland config %s does not exist", rootConfigPath)
-		}
-		return false, err
-	}
-
-	for _, sourceValue := range parseSourceLines(string(content)) {
-		sourcePaths, err := expandSourceValue(sourceValue, filepath.Dir(rootConfigPath))
-		if err != nil {
-			return false, err
-		}
-		for _, sourcePath := range sourcePaths {
-			if sourcePath == targetPath {
-				return true, nil
-			}
-
-			info, err := os.Stat(sourcePath)
-			if err != nil {
-				if errors.Is(err, os.ErrNotExist) {
-					continue
-				}
-				return false, err
-			}
-			if info.IsDir() {
-				continue
-			}
-			ok, err := isPathSourced(sourcePath, targetPath, visited)
-			if err != nil {
-				return false, err
-			}
-			if ok {
-				return true, nil
-			}
-		}
-	}
-	return false, nil
-}
-
-func parseSourceLines(content string) []string {
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	out := make([]string, 0)
-	for scanner.Scan() {
-		line := stripComments(scanner.Text())
-		line = strings.TrimSpace(line)
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, "=", 2)
-		if len(parts) != 2 {
-			continue
-		}
-		if strings.TrimSpace(parts[0]) != "source" {
-			continue
-		}
-		value := strings.TrimSpace(parts[1])
-		value = strings.Trim(value, `"'`)
-		if value != "" {
-			out = append(out, value)
-		}
-	}
-	return out
-}
-
-func stripComments(line string) string {
-	if idx := strings.Index(line, "#"); idx >= 0 {
-		return line[:idx]
-	}
-	return line
-}
-
-func expandSourceValue(value string, baseDir string) ([]string, error) {
-	resolved, err := resolvePath(value, baseDir)
-	if err != nil {
-		return nil, err
-	}
-	if hasGlob(resolved) {
-		matches, err := filepath.Glob(resolved)
-		if err != nil {
-			return nil, err
-		}
-		if len(matches) == 0 {
-			return []string{resolved}, nil
-		}
-		return matches, nil
-	}
-	return []string{resolved}, nil
 }
 
 func resolvePath(value string, baseDir string) (string, error) {
